add libressl-dev as build dependency
Resolves error:

    wget: can't execute 'ssl_helper': No such file or directory
    wget: error getting response: Connection reset by peer

Why did I switch to curl? 3.4 main repo is compiled against openssl,
but edge is compiled against libressl. That means there are several
options to resolve the error. In the end I chose one that does not
put "openssl" or "libressl" in the source code to avoid questions.

Neither are all that important to the build.
One of them is needed to fetch the github files.